REFACTOR: Pass metrics as a class parameter
Previously, MetricSpace classes used a self.distance static method as
the metric, which meant each metric defined on the underlying space
required a new class. Now, the class defines only the underlying space,
and the metric is passed as an argument on creation. Moreover, the
metrics have been moved to a module.